+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 - Style of the button componenets to have background on hover and only an outline otherwise.
 
+### Removed
+
+- Unused colours from the theme.
+
 ## [0.10.1] - 2021-04-12
 
 ### Fixed